requirements.txt
ipykernel==6.29.4
ipython==8.26.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.3
virtualenv==20.13.0+ds
python-socketio==5.11.4
python-engineio==4.9.1
tensorflow==2.17.0
scikeras==0.13.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4